Continuous univariate prior distribution

ContinuousPrior is a generic class for representing a univariate prior over a compact interval of the real numbers.

ContinuousPrior(pdf, support)

# S4 method for ContinuousPrior
bounds(dist, ...)

# S4 method for ContinuousPrior,`function`
expectation(dist, f,
  rel.tol = 0.001, ...)

# S4 method for ContinuousPrior,numeric
condition(dist, interval, ...)

# S4 method for DataDistribution,ContinuousPrior,numeric
predictive_pdf(dist,
  prior, x1, n1, k = 33, ...)

# S4 method for DataDistribution,ContinuousPrior,numeric
predictive_cdf(dist,
  prior, x1, n1, k = 33, ...)

# S4 method for DataDistribution,ContinuousPrior,numeric
posterior(dist, prior,
  x1, n1, ...)

Arguments

pdf
cf. slot pdf

support
cf. slot support

dist
data distribution or prior distribution (depending on context) [TODO: make this clearer by renaming arguments in generics!]

...
further optional arguments

f
univariate function of parameter to get expected value of

rel.tol
relative tolerance used in adaptive gaussian quadrature to evaluate the integral

interval
numeric vector of length two specifying the interval to condition on.

prior
prior distribution

x1
stage one outcome (note that n1 is available from design)

n1
stage-one sample size

k
number of pivots for crude integral approximation [TODO: this needs to be done properly!, cant use integrate since we wnat this vectorized!]

Details

For further details cf. Prior-class.

Functions

ContinuousPrior: constructor

Slots

pdf

function the actual probability density function , only called in the interior of support therefore need not be set to zero outside of support. [TODO: expose via pdf(prior, theta)!]

support

numeric vector of length two with the bounds of the compact interval on which the pdf is positive. The restriction to compact support simplifies numerical integration.