Data distributions

DataDistribution is an abstract class used to represent the distribution of a sufficient statistic x given a sample size n and a single parameter value theta.

# S4 method for DataDistribution
quantile(x, probs, n, theta, ...)

probability_density_function(dist, x, n, theta, ...)

  # S4 method for DataDistribution,numeric,numeric,numeric
probability_density_function(dist,
  x, n, theta, ...)

cumulative_distribution_function(dist, x, n, theta, ...)

  # S4 method for DataDistribution,numeric,numeric,numeric
cumulative_distribution_function(dist,
  x, n, theta, ...)

Arguments

x
outcome

probs
numeric vector of probabilities

n
sample size

theta
distribution parameter

...
further optional arguments

dist
data distribution object

Details

This abstraction layer allows representation of t-distributions (unknown variance), normal distribution (known variance), and normal approximation of a binary endpoint. Currently, only the normal case is implemented with Normal-class.

Methods (by generic)

quantile: quantile function of the respective distribution.

probability_density_function: probability density function given outcome, sample size and parameter; must be implemented.

cumulative_distribution_function: cumulative distribution function given outcome, sample size and parameter; must be implemented.