Two-stage design

Design is a abstract class for representing two-stage designs.

early_stopping_bounds(d, ...)

n1(d, ...)

n2(d, x1, ...)

n(d, x1, ...)

c2(d, x1, ...)

# S4 method for Design
early_stopping_bounds(d, ...)

# S4 method for Design
n1(d, ...)

# S4 method for Design,numeric
n2(d, x1, ...)

# S4 method for Design,numeric
c2(d, x1, ...)

# S4 method for Design,numeric
n(d, x1, ...)

# S4 method for Design
as.numeric(x)

# S4 method for Design
update(object)

Arguments

d
Design object

...
further optional arguments

x1
stage-one outcome

x
Design object to conver to numeric vector of design parameters.

object
Design object to update with new design parameter values

Details

Current main implementation: GQDesign-class [TODO add some details on two-stage designs]

Methods (by generic)

early_stopping_bounds: must return numeric vector of length two giving early stopping for futility / efficacy boundaries; must be finite. Tries to access d@c1f and d@c1e otherwise throws an error. This means that the method need not be reimplemented as long as the subclass contains fields c1f and c2e.

n1: must return stage-one sample size of design d; default implementation tries to access d@n1 otherwise throws an error.

n2: stage-two sample size given stage one-outcome, must be implemented.

c2: stage-two critical value given stage-one outcome, must be implemented.

n: overall sample size given stage-one outcome

as.numeric: convert design object to numeric vector of design parameters, must be implemented.

update: convert vector of design parameters to design object of the same class as object; inverse to as.numeric, must be implemented.