Quadratic smoothness penalty term

Smoothness_n2 is a generic class for implementing a smoothness penalty via the average squared first derivative of the stage two sample size function n2 of a two-stage design. The only parameter is the width used for the finite differences, h. The generic implementation only evluates n2 in the interior of the continuation region of a design.

Smoothness_n2(h = sqrt(.Machine$double.eps))

# S4 method for Smoothness_n2,Design
evaluate(s, design, specific = TRUE,
  ...)

Arguments

h
positive number, see slot h

s
an object of class Smoothness_n2

design
the design to compute the smoothness term for

specific
logical, should a design-specific implementation be used? defaults to TRUE. If TRUE, looks for design-specific .evaluate method and calls it

...
further optional arguments

Slots

h

positive number giving the width of the central finite difference interval for approximating the first derivative.